requirements.txt
bottle
six
PyJWT
pytest
coverage
coveralls
pytest-cov